refactor nixl tests
