+++ test/srt/run_suite.py
@@ -157,6 +157,9 @@ suites = {
         TestFile("test_disaggregation_pp.py", 140),
         TestFile("test_deepseek_v3_basic.py", 275),
         TestFile("test_deepseek_v3_mtp.py", 275),
+        TestFile("test_disaggregation_variable_parallelism_nixl.py", 1174),
+        TestFile("test_disaggregation_uniform_parallelism_nixl.py", 1174),
+
     ],
     "per-commit-4-gpu-b200": [
         # TestFile("test_gpt_oss_4gpu.py", 600),
